Add tests, fix feed bug

+++ src/main/java/net/ignaszak/socialnetwork/controller/rest/CommentsRestController.java
@@ -2,6 +2,7 @@ package net.ignaszak.socialnetwork.controller.rest;
 
 import net.ignaszak.socialnetwork.domain.Comment;
 import net.ignaszak.socialnetwork.domain.User;
+import net.ignaszak.socialnetwork.exception.ResourceNotFoundException;
 import net.ignaszak.socialnetwork.service.comment.CommentService;
 import net.ignaszak.socialnetwork.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,6 +29,7 @@ public class CommentsRestController {
     public void delete(@PathVariable Integer id) {
         User currentUser = userService.getCurrentUser();
         Comment comment = commentService.getCommentById(id);
+        if (comment == null) throw new ResourceNotFoundException();
         if (comment.isAuthor(currentUser) || comment.getPost().isAuthor(currentUser))
             commentService.delete(comment);
     }
